+++ src/main.rs
@@ -2,64 +2,61 @@ mod blockcopy;
 mod copier;
 mod messages;
 mod nullstream;
+mod foldersignature;
+mod ic;
+mod options;
+mod gui;
 
-use nullstream::NullStream;
-use messages::{Shared, Status};
+use messages::{Shared};
 use std::sync::{Mutex, Arc};
 use std::thread;
-use copier::Copier;
-use blockcopy::BufferBlockCopy;
 use std::time::Duration;
-use std::ops::DerefMut;
-use crate::messages::Command;
-use crossterm::event::{read, poll, Event, KeyCode};
+
+use ic::IC;
+
+use gui::GUI;
+use crate::foldersignature::FolderSignature;
+use crate::options::Options;
 
 fn main() {
 
-    //Shared data between the copier and the main thread
     let shared = Shared {
-        command: None,
-        status: Status::new(0, false),
+        gui_command: None,
+        status: Default::default(),
+        manager_command: None
     };
 
-    let main_shared = Arc::new(Mutex::new(shared));
+    let shared = Arc::from(Mutex::from(shared));
 
-    let mut copier = copier::Copier::new(Arc::clone(&main_shared));
+    let fs = FolderSignature::new("E:\\Software Projects\\IntelliJ\\ic\\source");
 
-    thread::spawn(move || copier.copy(& mut NullStream, & mut NullStream, &BufferBlockCopy));
+    let overall_size = fs.overall_size();
+
+    let gui = GUI::new(Arc::clone(&shared), overall_size);
+
+    let mut ic = IC::new(
+        "E:\\Software Projects\\IntelliJ\\ic\\destination",
+        fs,
+    Options,
+    Arc::clone(&shared));
+
+
+    thread::spawn(move || {
+        ic.resume().unwrap();
+    });
 
     loop {
 
-        let command = if let Ok(true) = poll(Duration::from_secs(0)) {
-            match read().unwrap() {
-                Event::Key(event) => {
-                    match event.code {
-                        KeyCode::Char('p') => {Some(Command::Pause)},
-                        KeyCode::Char('s') => {Some(Command::Play)},
-                        KeyCode::Char('q') => {Some(Command::Stop)},
+        let overall_progress = gui.update();
 
-                        _ => {None}
-                    }
-                }
-                _ => {None}
-            }
-        } else {
-            None
-        };
+        if overall_progress == overall_size {
+            gui.finish();
+            break;
+        }
 
-        let status = {
-            let mut lock = main_shared.lock();
-            let mut shared = lock.as_mut().unwrap().deref_mut();
+        thread::sleep(Duration::from_millis(100));
 
-            shared.command = command;
-
-            shared.status.clone()
-
-        };
-
-        println!("progress: {} {} {}", status.progress, status.paused, Arc::strong_count(&main_shared));
-
-        thread::sleep(Duration::from_millis(1000));
     }
+
 
 }